backend/image-service/requirements.txt
fastapi
uvicorn
python-multipart
cryptography
Pillow
img2pdf
